requirements.txt
certifi==2020.12.5
Faker==8.1.2
kafka-python==2.0.2
psycopg2==2.8.6
python-dateutil==2.8.1
six==1.16.0
text-unidecode==1.3
